Tweaked and updated the book loading system to write to file correctly using Qt's DataStream io in a attempt to help porting to Andiord (it didn't). I have now implemented QStandardDirectory, which dictates the correct locations for configs and data files should be written and read from. This change has allowed the Andiord app to now work and correctly load and save book and settings JSON's.
Upcoming changes will change how the two "saving" operations are implemented: copying from ui to book pointer values and writing book pointer values to actual JSON. The new behaviour will copy ui values to book pointer whenever there is a change in the ui values and writing to JSON occurs when the "Save" Action is triggered. This opens the door to more stable book saving and, in the future, allow changes to open books to have a "diff" operation applied.