Build and Documentation Improvements
Add descriptions to specific variables in the book data structures. Updated the settings ui to include units on a seperate column.  Restructured the mainwindow ui to be more space efficent. Fixed a issue with the settings dialog not reseting the changed settings on "reset." Chnaged the formating of confirmation books and other small details.